+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [PR #1120](https://github.com/rqlite/rqlite/pull/1120): go mod updates.
 - [PR #1126](https://github.com/rqlite/rqlite/pull/1126): Add Queue instrumentation.
 - [PR #1127](https://github.com/rqlite/rqlite/pull/1127): Clearer Queued Writes loop logic.
+- [PR #1128](https://github.com/rqlite/rqlite/pull/1128): Add ResetStats to some modules.
 
 ## 7.11.0 (November 15th 2022)
 ### New features